Conflicts:
	Makefile
	NuttX
	cmake/configs/nuttx_px4fmu-v2_default.cmake
	nuttx-configs/aerofc-v1/nsh/defconfig
	nuttx-configs/auav-x21/include/nsh_romfsimg.h
	nuttx-configs/crazyflie/nsh/Make.defs
	nuttx-configs/crazyflie/nsh/defconfig
	nuttx-configs/mindpx-v2/nsh/defconfig
	src/drivers/boards/aerofc-v1/CMakeLists.txt
	src/drivers/boards/aerofc-v1/aerofc_init.c
	src/drivers/boards/aerofc-v1/board_config.h
	src/drivers/boards/asc-v1/asc_spi.c
	src/drivers/boards/crazyflie/CMakeLists.txt
	src/drivers/boards/crazyflie/board_config.h
	src/drivers/boards/crazyflie/crazyflie_spi.c
	src/drivers/boards/mindpx-v2/mindpx2_init.c
	src/drivers/boards/mindpx-v2/mindpx_spi.c
	src/drivers/boards/tap-v1/board_config.h
	src/drivers/drv_gpio.h
	src/firmware/nuttx/CMakeLists.txt
	src/modules/gpio_led/gpio_led.c
	src/modules/uavcanesc/commands/cfg/esc_cfg.hpp
	src/platforms/px4_defines.h

+++ src/modules/mavlink/mavlink_log_handler.h
@@ -40,6 +40,7 @@
 #include <queue.h>
 #include <time.h>
 #include <stdio.h>
+#include <cstdbool>
 #include <v2.0/mavlink_types.h>
 #include "mavlink_stream.h"
 
@@ -99,7 +100,7 @@ public:
 
 	// Overrides from MavlinkStream
 	const char     *get_name(void) const;
-	uint8_t         get_id(void);
+	uint16_t        get_id(void);
 	unsigned        get_size(void);
 	void            send(const hrt_abstime t);
